refactor: streamline entry action handling and replace ActionButton with CommandActionButton
- Removed unused imports and simplified the use of commands in `useEntryActions`.
- Updated `EntryItemWrapper`, `HeaderActions`, and `MoreActions` components to utilize `CommandActionButton` for better command handling.
- Introduced `CommandDropdownMenuItem` for consistent command representation in dropdowns.
- Enhanced readability and maintainability by reducing redundant code and improving type safety.
