+++ apps/renderer/src/modules/entry-content/header.mobile.tsx
@@ -1,4 +1,4 @@
-import { ActionButton, MotionButtonBase } from "@follow/components/ui/button/index.js"
+import { MotionButtonBase } from "@follow/components/ui/button/index.js"
 import { RootPortal } from "@follow/components/ui/portal/index.js"
 import { findElementInShadowDOM } from "@follow/utils/dom"
 import { clsx, cn } from "@follow/utils/utils"
@@ -10,6 +10,7 @@ import { useEventCallback } from "usehooks-ts"
 
 import { useUISettingKey } from "~/atoms/settings/ui"
 import { HeaderTopReturnBackButton } from "~/components/mobile/button"
+import { CommandActionButton } from "~/components/ui/button/command-button"
 import { useScrollTracking, useTocItems } from "~/components/ui/markdown/components/hooks"
 import { ENTRY_CONTENT_RENDER_CONTAINER_ID } from "~/constants/dom"
 import type { EntryActionItem } from "~/hooks/biz/useEntryActions"
@@ -17,6 +18,8 @@ import { useEntryActions } from "~/hooks/biz/useEntryActions"
 import { useEntry } from "~/store/entry/hooks"
 
 import { COMMAND_ID } from "../command/commands/id"
+import { useCommand } from "../command/hooks/use-command"
+import type { FollowCommandId } from "../command/types"
 import { useEntryContentScrollToTop, useEntryTitleMeta } from "./atoms"
 import type { EntryHeaderProps } from "./header.shared"
 
@@ -87,13 +90,12 @@ function EntryHeaderImpl({ view, entryId, className }: EntryHeaderProps) {
           )}
         >
           {actionConfigs.map((item) => (
-            <ActionButton
-              icon={item.icon}
+            <CommandActionButton
+              key={item.id}
+              commandId={item.id}
+              onClick={item.onClick}
               active={item.active}
               shortcut={item.shortcut}
-              onClick={item.onClick}
-              tooltip={item.name}
-              key={item.name}
             />
           ))}
         </div>
@@ -177,18 +179,14 @@ const HeaderRightActions = ({
                 >
                   <div className="flex flex-col items-center py-2">
                     {actions.map((item) => (
-                      <MotionButtonBase
+                      <CommandMotionButton
+                        key={item.id}
+                        commandId={item.id}
                         onClick={() => {
                           setCtxOpen(false)
                           item.onClick?.()
                         }}
-                        key={item.name}
-                        layout={false}
-                        className="flex w-full items-center gap-2 px-4 py-2"
-                      >
-                        {item.icon}
-                        {item.name}
-                      </MotionButtonBase>
+                      />
                     ))}
                   </div>
                 </m.div>
@@ -200,6 +198,28 @@ const HeaderRightActions = ({
     </div>
   )
 }
+
+const CommandMotionButton = ({
+  commandId,
+  onClick,
+}: {
+  commandId: FollowCommandId
+  onClick: () => void
+}) => {
+  const command = useCommand(commandId)
+  if (!command) return null
+  return (
+    <MotionButtonBase
+      onClick={onClick}
+      layout={false}
+      className="flex w-full items-center gap-2 px-4 py-2"
+    >
+      {command.icon}
+      {command.label.title}
+    </MotionButtonBase>
+  )
+}
+
 const TableOfContentsIcon = ({ className = "size-6" }) => {
   return (
     <svg
